requirements.txt
requests
python-dotenv
json
uuid
websocket
googletrans
Pillow